Inline DQMEventInfo.h.
